% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaRobust.R
\name{pcaRobust}
\alias{pcaRobust}
\title{pcaRobust}
\usage{
pcaRobust()
}
\description{
calculate principal component analysis Robust with ROBPCA method
}
\examples{
library(Rcmdr)
#pcaRobust()
}

